requirements.txt
configurations==0.0.1
Django==4.0.3
django_configurations==2.3.2
django_versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
Pillow==9.2.0
